+++ src/app/(main)/(private)/payments/success/page.tsx
@@ -1,35 +1,12 @@
-"use client";
-
-import { useEffect, useRef } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { useApprovePayments } from "@/features/payments/hooks/useApprovePayments";
+import Success from "@/components/payments/Success";
+import { Suspense } from "react";
 
 export default function PaymentsSuccessPage() {
-  const params = useSearchParams();
-  const router = useRouter();
-  const approvePayments = useApprovePayments();
-
-  const paymentKey = params.get("paymentKey");
-  const orderId = params.get("orderId");
-  const amount = Number(params.get("amount"));
-
-  const calledRef = useRef(false);
-
-  useEffect(() => {
-    if (calledRef.current) return;
-    calledRef.current = true;
-
-    if (!paymentKey || !orderId || !Number.isFinite(amount)) {
-      router.push("/mypage/bizz");
-      return;
-    }
-
-    approvePayments.mutate({ paymentKey, orderId, amount });
-  }, [paymentKey, orderId, amount, approvePayments, router]);
-
   return (
-    <div className="flex min-h-screen items-center justify-center">
-      <p className="text-lg font-semibold">결제 승인 처리 중입니다...</p>
-    </div>
+    <>
+      <Suspense fallback={null}>
+        <Success />
+      </Suspense>
+    </>
   );
 }
